refactor: 콘솔 controller 추가 및 기능 분리
